Portfolio Streamlit Skeleton
This is the basic layout of the portfolio. It will undergo changes to the background, images, font size, and coloring for certain elements through a css style sheet. There will be an inclusion of a navigation bar and a working contact me page that will send emails directly to me email.